requirements.txt
alembic==1.7.7
aniso8601==9.0.1
attrs==21.4.0
bcrypt==3.2.2
bidict==0.22.0
blinker==1.4
cffi==1.15.0
click==8.1.3
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
flask-expects-json==1.7.0
Flask-JWT-Extended==4.4.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
h11==0.13.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.5.1
Mako==1.2.0
MarkupSafe==2.1.1
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.0.2
pyrsistent==0.18.1
python-engineio==4.3.4
python-socketio==5.7.1
pytz==2022.1
simple-websocket==0.8.0
six==1.16.0
SQLAlchemy==1.4.36
websockets==10.3
Werkzeug==2.1.2
wsproto==1.1.0
